Use macro for FROST types in node bindings.
