+++ crates/bindings/node/src/protocols/frost/core.rs
@@ -95,4 +95,127 @@ macro_rules! frost_impl {
     };
 }
 
+macro_rules! frost_types {
+    () => {
+        impl TryFrom<ThresholdKeyShare> for KeyShare {
+            type Error = polysig_protocol::Error;
+
+            fn try_from(
+                value: ThresholdKeyShare,
+            ) -> std::result::Result<Self, Self::Error> {
+                let key_share: driver::KeyShare =
+                    (&value).try_into()?;
+                Ok(key_share.into())
+            }
+        }
+
+        impl TryFrom<KeyShare> for ThresholdKeyShare {
+            type Error = polysig_protocol::Error;
+
+            fn try_from(
+                value: KeyShare,
+            ) -> std::result::Result<Self, Self::Error> {
+                let key_share: driver::KeyShare = value.into();
+                Ok((&key_share).try_into()?)
+            }
+        }
+
+        #[doc(hidden)]
+        #[napi(object)]
+        #[derive(Serialize, Deserialize, Debug)]
+        #[serde(rename_all = "camelCase")]
+        pub struct Signature {
+            pub signature_bytes: Vec<u8>,
+        }
+
+        impl TryFrom<Signature> for frost::Signature {
+            type Error = napi::Error;
+
+            fn try_from(
+                value: Signature,
+            ) -> std::result::Result<Self, Self::Error> {
+                let signature = frost::Signature::deserialize(
+                    &value.signature_bytes,
+                )
+                .map_err(Error::new)?;
+                Ok(signature)
+            }
+        }
+
+        impl TryFrom<frost::Signature> for Signature {
+            type Error = napi::Error;
+
+            fn try_from(
+                value: frost::Signature,
+            ) -> std::result::Result<Self, Self::Error> {
+                let signature_bytes =
+                    value.serialize().map_err(Error::new)?;
+                Ok(Self { signature_bytes })
+            }
+        }
+
+        #[doc(hidden)]
+        #[napi(object)]
+        #[derive(Serialize, Deserialize, Debug)]
+        #[serde(rename_all = "camelCase")]
+        pub struct Identifier {
+            pub identifier_bytes: Vec<u8>,
+        }
+
+        impl TryFrom<Identifier> for frost::Identifier {
+            type Error = napi::Error;
+
+            fn try_from(
+                value: Identifier,
+            ) -> std::result::Result<Self, Self::Error> {
+                let identifier = frost::Identifier::deserialize(
+                    &value.identifier_bytes,
+                )
+                .map_err(Error::new)?;
+                Ok(identifier)
+            }
+        }
+
+        impl From<frost::Identifier> for Identifier {
+            fn from(value: frost::Identifier) -> Self {
+                let identifier_bytes = value.serialize();
+                Self { identifier_bytes }
+            }
+        }
+
+        #[doc(hidden)]
+        #[napi(object)]
+        #[derive(Debug, Serialize, Deserialize)]
+        #[serde(rename_all = "camelCase")]
+        pub struct PartyOptions {
+            pub public_key: Vec<u8>,
+            pub participants: Vec<Vec<u8>>,
+            pub is_initiator: bool,
+            pub verifiers: Vec<VerifyingKey>,
+        }
+
+        impl TryFrom<PartyOptions> for frost::PartyOptions {
+            type Error = napi::Error;
+
+            fn try_from(
+                value: PartyOptions,
+            ) -> std::result::Result<Self, Self::Error> {
+                let mut verifiers =
+                    Vec::with_capacity(value.verifiers.len());
+                for verifier in value.verifiers {
+                    verifiers.push(verifier.try_into()?);
+                }
+                Ok(polysig_driver::PartyOptions::new(
+                    value.public_key,
+                    value.participants,
+                    value.is_initiator,
+                    verifiers,
+                )
+                .map_err(Error::new)?)
+            }
+        }
+    };
+}
+
 pub(crate) use frost_impl;
+pub(crate) use frost_types;
